fix(grid): Change grid generation logic
Previously, the game would not work properly most of the time since
there was a logical error when generating the grid, making it so that
the populateShips method would fall into an endless loop. After the fix,
the grid is properly generated and the program can now detect the
position of the mouse.